Updated version etc.
Updated from 1.1 to 1.1.1 and updated date.  This can all be altered
again pre-release, my intent was to be able to distinguish our
development version from the CRAN version when installing in R.

Removed boot from Imports (we're going to build our own bootstrap
routine in abund.estim and likely the kernel code) and added SparseM
(used in kernel code).