Updates to vignette.example
The example now loads example data, fits a dfunc, and estimates
abundance with bootstrap-based CIs.
Issues:  The CI doesn't contain the estimate and may not be printing all
the output it needs to.  Will revisit and revise.  Noticed that n.hat
was being calculated with two different formulas within F.abund.estim
(one way for the estimate, and a different equation for inside the old
bootstrap loop).  Will look into ...