Bug fix in F.abund.estim
I noticed the C.I. for N was too low (not containing n.hat), and found
the bug in my code as to why (even if a transect was resampled >1 time
in the bootstrap sample, the counts associated with that transect were
only represented once -- thus the bootstrap estimates of n.hat were
biased low).  Fixed!