Replace anyNA function
Fixing an error found by CRAN check on older R version.  These three
functions used the anyNA function to look for NAs and raise an error
message if found.  Apparently, anyNA is only in R versions > 3.1.0, so
we're replacing with base functionality to make compatible with these
older versions as well.