Updates to help documentation
The changes to datasets and functions warranted changes to the many of
the help documents -- mainly the example section, but also some argument
names changed, etc.

Also made some rewrites to the package's help doc (Rdistance-package.Rd)
to update and make complementary to vignettes and material that will be
on the GitHub site/wiki.