Add utils to Imports
utils is a base package,  so we're not requiring an additional package
be installed.  It is very hand to have utils for the packageVersion
function that I've implemented in the zzz.R file and in the vignettes in
order to automate the updating of the version number to match the
DESCRIPTION file.