Changed F.abund.estim to check for NA's only in key columns of data frames.
Prior to this, Jason (that newbe), was checking for NA's anywhere in the data frames.  I hope I'm not missing something, but I think NA's are okay as long as they do not occur in the key columns like dist, siteID, length, and groupsize.

I also changed zzz.r to automatically update the version number.