Major update in F.abund.estim
F.abund.estim appears to work with both lines and points (except see
note on by.id argument below), when there are no covariates.  Still need
to test and debug an example with covariates.  The by.id argument still
needs to be updated for points and for covariates.  I added a function
called estimate.nhat, because we were copy/pasting this a few times.
Should probably be split out to new estimate.nhat.R script.