Added equation to estimate abundance at points with no covars
The "general" method Aidan and Trent added was being applied to line
data with covars and all point data.  If no covars, there is a simpler
method for abundance estimation at points (the one for lines was already
in).  Also made some notes, because this new equation matches Distance,
but not our general method.