Bug fix in extracting formula
The detection function formula was being extracted from the call.  This works when the user specified the forumla explicitly (dist ~ shrub), but failed when the formula was created using a paste function (e.g., paste0("dist", "~", "bare")).  This paste functionality is key to a new autoDistSamp Diem is working on that fits different covariates.  But the formula was already being stored in the dfunc object, so using that instead.  dfunc$formula works with paste.