Changes for unitless parameters and preventing return of 0
Before: we assumed sigma of halfnorm had units. For extremely bad guesses at sigma, integration.constant returned 0, which resulted in NaN's, which stuck the optimization. 

After: Removed drop_units from sigma.  And, we return .Machine$double.xmax / (nzeros) when integration constant is zero.  That is, we want to return a huge integration constant when guess is terrible so that log(-like) is huge but not NaN

files committed: 
  R/integration.constant.R