Update of predict method
I thought this was already committed, and am not sure anything has changed, but git says so.

Before: users has to specify full factors, with levels, in the newdata matrix.

After: users can specify levels of a factor using strings in newdata, similar to the way you do in lm()

files committed:
  R/predict.dfunc.R