Bug fix for Issue 107. Fixed scaling of covariate dfuncs
Before: scaling of predicted distance functions was messed up. Inside predict.dfunc, y matrix was wrong way around and needed to be transposed in order for the scaling to work.

After: y is transposed and scaler vector is properly applied. Made significant additions to help for predict.dfunc and added examples.

Files committed:
  R/predict.dfunc.R
  man/predict.dfunc.Rd